Phase 10: Nintendo 3DS backend (code-complete, toolchain-verified, not hardware-verified)
Implements the final phase per the project's 10-phase plan: real IRenderer/
InputState-feed/IAudioBackend implementations targeting the 3DS, plus a real
entry point, all cross-compiled and linked against a from-source-bootstrapped
devkitARM-equivalent toolchain (devkitPro's own servers are unreachable from
this build environment; GitHub is). See docs/3ds-toolchain.md for the full
bootstrap writeup.

New 3DS backend code (src/renderer/Nintendo3DSRenderer.*, src/platform/
Nintendo3DSInput.*, src/audio/Nintendo3DSAudioBackend.*, src/platform/
nintendo3ds_main.cpp):
- Nintendo3DSRenderer wraps a SoftwareRenderer by composition and blits its
  finished RGBA8 framebuffer to the real 3DS LCD framebuffer via libctru's
  gfxGetFramebuffer(), using a rotated/column-major indexing formula
  cross-checked against libctru's own source/gfx.c (not assumed). Top screen
  only; bottom-screen use is an explicit follow-up.
- Nintendo3DSInput polls libctru hid once per frame and feeds InputState:
  touch->mouse, D-Pad/Circle Pad->arrow keys, face buttons/START/SELECT->
  reasonable-effort Key.* stand-ins (documented, not spec-derived).
- Nintendo3DSAudioBackend does real ndsp channel reservation/pause/stop
  plumbing, but -- like NullAudioBackend before it -- still can't play any
  actual audio, since no SWF sound codec decode exists anywhere in this
  codebase yet (documented explicitly, not silently narrowed).
- nintendo3ds_main.cpp ties it all together into a real app loop, playing an
  embedded demo movie (SD-card loading is an explicit, well-scoped follow-up
  -- needs FSUSER calls or restoring the excluded archive_dev.c).

Embedded demo content: tools/gen_3ds_demo_swf.py is a from-scratch,
clean-room SWF byte-level generator (hand-authored against the public SWF
spec, shares no code with tests/SwfTestFixtures.cpp) producing a small demo
movie (a red square sliding across a blue background over 8 frames),
committed as the generated src/platform/EmbeddedDemoSwf.h. Independently
verified via this project's own desktop flash_runtime CLI (--timeline,
--render, pixel-sampled) before embedding -- found and fixed two bugs in the
generator itself along the way: a bogus HasTranslate bit in the hand-rolled
MATRIX encoder (the real spec has no such flag) and a wrong PlaceObject2
HasMatrix bit assignment (0x08 instead of 0x04, cross-checked against this
project's own PlaceObjectTag.cpp parser).

Toolchain bootstrap (docs/3ds-toolchain.md, cmake/Toolchain-3DS.cmake,
third_party/3ds-support/ -- MPL-2.0 devkitarm-crtls files vendored verbatim,
see that directory's README.md for provenance): Ubuntu's generic
gcc-arm-none-eabi cross-compiler plus libctru/citro3d/devkitarm-crtls built
directly from their public GitHub sources, targeting ARMv6K/MPCore (a strict
ISA superset of the only multilib Ubuntu ships, ARMv5TE, with an identical
EABI hard-float calling convention). Validated end-to-end in this session:
flash3ds_core (completely unmodified) plus the new 3DS code cross-compiles
and links with zero undefined non-weak symbols into a structurally valid
.3dsx (correct magic bytes, confirmed via od). Every build/link issue hit
during the bootstrap and its fix is documented in full, including: the
ARMv6K/-mfpu=vfp requirement, the missing sync-dmb.specs/sys/iosupport.h/
__SYSCALL macro/_LOCK_RECURSIVE_T newlib gaps (each resolved with a minimal,
justified stub/patch/exclusion -- see flash3ds_syscalls.c and
synchronization.h's patch comment for exactly what's reused from libctru's
own SDK vs. newly written), the -lcitro3d-before-lctru link-order
requirement, and the -specs=nosys.specs requirement for newlib's default
"always fail" syscall stubs.

Two genuine (if narrow) portability bugs in the platform-independent core
were found and fixed while bringing up the cross-compile -- uint32_t is not
the same type as unsigned int/int on this ARM target the way it happens to
be on x86_64 desktop, which had silently miscompiled two std::clamp/
std::min/std::max call sites:
- Timeline::gotoAndStop()/gotoAndPlay(): std::clamp(frameIndex, 1u, ...)
  now casts the literal to uint32_t explicitly.
- SoftwareRenderer::fillPolygon(): minY/maxY (int) vs. PointTwips::y
  (int32_t) now cast explicitly at each comparison.
Both fixes are platform-independent and benefit every target, not 3DS-only
workarounds. All 187 desktop tests still pass after every change in this
phase (rebuilt and re-verified repeatedly, per the project's own
phase-completion workflow).

IMPORTANT CAVEAT (documented prominently in docs/3ds-toolchain.md's "What's
verified vs. not" and mirrored in CLAUDE.md): this environment has no 3DS
hardware or emulator available. The .3dsx has never actually been booted or
had its rendered output visually confirmed. Phase 10 is code-complete and
toolchain-verified, not hardware-verified -- that's the natural next step
for a session with access to real hardware or Citra.

docs/ updated: 3ds-toolchain.md (new), architecture.md, renderer.md,
audio.md, input.md, swf-support.md, avm1-support.md, CLAUDE.md -- all with
Phase 10 sections matching this project's existing documentation style and
confidence-flagging conventions.

+++ src/renderer/SoftwareRenderer.cpp
@@ -61,11 +61,16 @@ void SoftwareRenderer::fillPolygon(const std::vector<PointTwips>& devicePoints,
                                     swf::RgbaColor color) {
     if (devicePoints.size() < 3) return;
 
-    int minY = devicePoints[0].y;
-    int maxY = devicePoints[0].y;
+    int minY = static_cast<int>(devicePoints[0].y);
+    int maxY = static_cast<int>(devicePoints[0].y);
     for (const auto& p : devicePoints) {
-        minY = std::min(minY, p.y);
-        maxY = std::max(maxY, p.y);
+        // Explicit cast: PointTwips::y is int32_t, which is NOT guaranteed
+        // to be the same type as `int` (they coincide on x86_64 desktop
+        // builds but not on this project's ARM/newlib cross-compile
+        // target -- see Timeline.cpp's gotoAndStop() for the same class of
+        // portability issue with std::clamp).
+        minY = std::min(minY, static_cast<int>(p.y));
+        maxY = std::max(maxY, static_cast<int>(p.y));
     }
     minY = std::max(minY, 0);
     maxY = std::min(maxY, height_ - 1);
